Distance coefficients should be doubles

+++ src/lib/CameraSource.h
@@ -30,12 +30,12 @@ public:
     int getHeight() const { return mHeight; }
     GLuint getTexture(Eye e) const { return mTexture[e]; }
     cv::Mat& getCaneraMatrix() { return mCameraMatrix; }
-    std::vector<float>& getDistCoeffs() { return mDistCoeffs; }
+    std::vector<double>& getDistCoeffs() { return mDistCoeffs; }
 
 protected:
     int mWidth, mHeight;
     GLuint mTexture[2];
     cv::Mat mCameraMatrix;
-    std::vector<float> mDistCoeffs;
+    std::vector<double> mDistCoeffs;
 
 };
